@@ -1,8 +1,8 @@
 import Vue from 'vue'
-import App from './App'
+import App from './App.vue'
 
-/* eslint-disable no-new */
+Vue.config.productionTip = false
+
 new Vue({
-  el: 'app',
   render: h => h(App)
-})
+}).$mount('#app')
